requirements.txt
streamlit>=1.35.0
openpyxl>=3.1.2
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
